fix: support UtcOffset as decimal (#70)

+++ src/Raygun.Blazor/Models/BrowserSpecs.cs
@@ -125,9 +125,9 @@ namespace Raygun.Blazor.Models
         public string? UserAgent { get; set; }
 
         /// <summary>
-        /// UTC offset in minutes. From `new Date().getTimezoneOffset() / -60`.
+        /// UTC offset in hours. From `new Date().getTimezoneOffset() / -60`.
         /// </summary>
-        public int UtcOffset { get; set; }
+        public decimal UtcOffset { get; set; }
 
         #endregion
